requirements.txt
astroid==2.2.4
certifi==2018.11.29
chardet==3.0.4
GroupyAPI==0.10.3
idna==2.7
isort==4.3.12
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.7
pylint==2.3.1
requests==2.20.1
six==1.12.0
typed-ast==1.3.1
urllib3==1.24.1
wrapt==1.11.1
